+++ packages/server/src/lib/measurement-value/binary.js
@@ -1,0 +1,9 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class Binary extends MeasurementValue {
+
+}
+
+module.exports = {
+  Binary,
+};
