+++ packages/server/src/lib/measurement-value/comment.js
@@ -1,0 +1,9 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class Comment extends MeasurementValue {
+
+}
+
+module.exports = {
+  Comment,
+};
