+++ packages/server/src/lib/measurement-value/float-point.js
@@ -1,0 +1,9 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class FloatPoint extends MeasurementValue {
+
+}
+
+module.exports = {
+  FloatPoint,
+};
