+++ packages/server/src/lib/measurement-value/index.js
@@ -1,0 +1,19 @@
+const { Binary } = require('./binary');
+const { Comment } = require('./comment');
+const { FloatCdf } = require('./float-cdf');
+const { FloatPoint } = require('./float-point');
+const { Null } = require('./null');
+const { Percentage } = require('./percentage');
+const { UnresolvableResolution } = require('./unresolvable-resolution');
+const { MeasurementValue } = require('./measurement-value');
+
+module.exports = {
+  Binary,
+  Comment,
+  FloatCdf,
+  FloatPoint,
+  Null,
+  Percentage,
+  UnresolvableResolution,
+  MeasurementValue,
+};
