+++ packages/server/src/lib/measurement-value/null.js
@@ -1,0 +1,11 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class Null extends MeasurementValue {
+  validate() {
+    return false;
+  }
+}
+
+module.exports = {
+  Null,
+};
