+++ packages/server/src/lib/measurement-value/percentage.js
@@ -1,0 +1,9 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class Percentage extends MeasurementValue {
+
+}
+
+module.exports = {
+  Percentage,
+};
