+++ packages/server/src/lib/measurement-value/unresolvable-resolution.js
@@ -1,0 +1,9 @@
+const { MeasurementValue } = require('./measurement-value');
+
+class UnresolvableResolution extends MeasurementValue {
+
+}
+
+module.exports = {
+  UnresolvableResolution,
+};
